+++ web_client/src/analytics/TrackedErrorBoundary/OuterFallbackErrorPage.tsx
@@ -1,0 +1,9 @@
+import { FallbackProps } from "react-error-boundary";
+import { FallbackErrorPage } from ".";
+
+export default function OuterFallbackErrorPage({
+  error,
+  resetErrorBoundary,
+}: FallbackProps) {
+  return <FallbackErrorPage error={error} reset={resetErrorBoundary} />;
+}
